+++ LinkNoteSpring/src/main/java/com/linknote/online/linknotespring/note/noteService/NoteServiceImpl.java
@@ -6,17 +6,13 @@ import com.linknote.online.linknotespring.note.notedto.GetTagsParamDto;
 import com.linknote.online.linknotespring.note.notedto.UpdateNoteTagParamDto;
 import com.linknote.online.linknotespring.note.notedto.DeleteNoteParamDto;
 import com.linknote.online.linknotespring.note.notedto.GetNoteParamDto;
-import com.linknote.online.linknotespring.note.notedto.GetNotesParamDto;
 import com.linknote.online.linknotespring.note.notedto.UpdateNoteParamsDto;
 import com.linknote.online.linknotespring.note.notedto.UpdateNoteSharedParamDto;
 import com.linknote.online.linknotespring.note.notedto.UpdateNoteStarParamDto;
 import com.linknote.online.linknotespring.note.noteexception.NoteAlreadyExistException;
 import com.linknote.online.linknotespring.note.notepo.po.NotePO;
-import com.linknote.online.linknotespring.note.notepo.po.NotesPO;
 import com.linknote.online.linknotespring.note.notepo.response.NoteResPO;
-import com.linknote.online.linknotespring.note.notepo.response.NotesResPO;
 import com.linknote.online.linknotespring.note.notepo.response.TagResPO;
-import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -44,12 +40,9 @@ public class NoteServiceImpl implements NoteService{
   }
 
   @Override
-  public void createNote(CreateNoteParamsDto params) {
-    Integer result = noteDao.getNoteIdByNameForVerifyNameExist(params.getNoteName(), params.getNotebookId());
-    if(result != null){
-      throw new NoteAlreadyExistException("NoteService: note name already exist ");
-    }
-    noteDao.createNote(params);
+  public Integer createNote(CreateNoteParamsDto params) {
+    params.setNoteName("new note");
+    return noteDao.createNote(params);
   }
 
   @Override
