+++ LinkNoteSpring/src/main/java/com/linknote/online/linknotespring/note/notedao/NoteDaoImpl.java
@@ -11,6 +11,7 @@ import com.linknote.online.linknotespring.note.notepo.po.NotePO;
 import com.linknote.online.linknotespring.note.notepo.po.NotesPO;
 import com.linknote.online.linknotespring.note.noterowmapper.NoteRowMapper;
 import com.linknote.online.linknotespring.note.noterowmapper.NotesRowMapper;
+import io.swagger.v3.oas.models.security.SecurityScheme.In;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.HashMap;
@@ -19,7 +20,10 @@ import java.util.Map;
 import java.util.Objects;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 
 @Repository
@@ -67,12 +71,14 @@ public class NoteDaoImpl implements NoteDao{
   }
 
   @Override
-  public void createNote(CreateNoteParamsDto params) {
-    String sql = "INSERT INTO notes (name, notebookId) VALUES (:noteName, :notebookId)";
+  public Integer createNote(CreateNoteParamsDto params) {
+    String sql = "INSERT INTO notes (name, notebookId) VALUES (:name, :notebookId)";
     Map<String, Object> map = new HashMap<>();
-    map.put("noteName", params.getNoteName());
     map.put("notebookId", params.getNotebookId());
-    namedParameterJdbcTemplate.update(sql,map);
+    map.put("name", params.getNoteName());
+    KeyHolder keyHolder = new GeneratedKeyHolder();
+    namedParameterJdbcTemplate.update(sql,new MapSqlParameterSource(map), keyHolder);
+    return Objects.requireNonNull(keyHolder.getKey()).intValue();
   }
 
   @Override
