+++ LinkNoteSpring/src/main/java/com/linknote/online/linknotespring/note/notedto/CreateNoteParamsDto.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public class CreateNoteParamsDto {
-  @NotBlank
+
   private String noteName;
 
   private Integer notebookId;
