--- md.html
@@ -1,74 +1,0 @@
-<!DOCTYPE html>
-<html lang="zh-TW">
-  <head>
-    <meta charset="UTF-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1, shrink-to-fit=no"
-    />
-    <title>自行輸入程式碼的高亮範例</title>
-
-    <!-- 引入 Bootstrap CSS -->
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-    />
-
-    <!-- 引入 Prism 的 CSS -->
-    <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/prism-themes@1.1.0/themes/prism-vs.css"
-    />
-  </head>
-  <body>
-    <!-- 輸入程式碼的 textarea -->
-    <div class="container mt-5">
-      <h1>自行輸入程式碼的高亮範例</h1>
-
-      <label for="codeInput" class="form-label">輸入程式碼：</label>
-      <textarea
-        class="form-control"
-        id="codeInput"
-        rows="10"
-        placeholder="在這裡輸入程式碼..."
-      ></textarea>
-
-      <div class="mt-3">
-        <h2>預覽：</h2>
-        <!-- 用來顯示高亮後的程式碼的區域 -->
-        <pre>
-            <code id="codeOutput" class="language-none">
-                <!-- 這裡將顯示高亮後的程式碼 -->
-            </code>
-        </pre>
-      </div>
-    </div>
-
-    <!-- 引入 Prism 的 JavaScript 和 Markdown 解析器 -->
-    <script src="https://cdn.jsdelivr.net/npm/prism-themes@1.1.0/prism.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/marked@2.1.3/marked.min.js"></script>
-
-    <!-- 使用 JavaScript 監聽 textarea 的輸入事件並即時更新高亮效果 -->
-    <script>
-      document.addEventListener("DOMContentLoaded", function () {
-        var codeInput = document.getElementById("codeInput");
-        var codeOutput = document.getElementById("codeOutput");
-
-        // 監聽 textarea 的輸入事件
-        codeInput.addEventListener("input", function () {
-          // 取得 textarea 中的程式碼文字
-          var codeText = this.value;
-
-          // 使用 Marked.js 將 Markdown 格式的程式碼轉換為 HTML
-          var htmlContent = marked("```\n" + codeText + "\n```");
-
-          // 將轉換後的 HTML 插入到指定的區域中
-          codeOutput.innerHTML = htmlContent;
-
-          // 初始化 Prism
-          Prism.highlightAll();
-        });
-      });
-    </script>
-  </body>
-</html>
