+++ static/base/style/notification.css
@@ -1,0 +1,30 @@
+.wrapper-notification {
+  position: fixed;
+  top: 20px;
+  right: 0;
+  color: white;
+}
+
+.notice {
+  padding: 10px;
+  border-top-left-radius: 10px;
+  border-bottom-left-radius: 10px;
+  text-align: center;
+  margin-top: 10px;
+  transition: all 0.5s ease-in-out;
+  -webkit-user-select: none;
+     -moz-user-select: none;
+          user-select: none;
+}
+
+#success {
+  background-color: green;
+}
+
+#failed {
+  background-color: red;
+}
+
+#warn {
+  background-color: #e9a218;
+}/*# sourceMappingURL=notification.css.map */
