--- static/notePage/main/writingArea/style/style.css
@@ -1,78 +1,0 @@
-.edit-ctn-noteContent {
-  background-color: rgba(22, 22, 22, 0.6);
-  border-radius: 10px;
-  position: relative;
-  margin-top: 30px;
-}
-.edit-ctn-noteContent .tagBtn {
-  position: relative;
-}
-.edit-ctn-noteContent .tagBtn img {
-  width: 25px;
-  height: 25px;
-  cursor: pointer;
-}
-.edit-ctn-noteContent .tagBtn .tagList {
-  position: absolute;
-}
-
-.noteContent-tools {
-  width: 270px;
-  background-color: #333333;
-  padding: 10px 20px;
-  border-radius: 10px;
-  gap: 10px;
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: -20px;
-  margin: auto;
-}
-.noteContent-tools img {
-  width: 25px;
-  height: 25px;
-  cursor: pointer;
-}
-.noteContent-tools div {
-  border-left: 2px solid #aaaaaa;
-  margin-inline: 8px;
-}
-
-.edit-itme-noteContent {
-  padding: 20px;
-  flex-direction: column;
-  gap: 15px;
-}
-.edit-itme-noteContent textarea {
-  display: block;
-  background-color: #222222;
-  border: none;
-  border-radius: 10px;
-  resize: none;
-  padding: 20px;
-  color: white;
-  line-height: 1.5;
-}
-.edit-itme-noteContent textarea::-webkit-scrollbar {
-  display: none;
-}
-.edit-itme-noteContent textarea:nth-child(2) {
-  height: calc(100vh - 423px);
-}
-.edit-itme-noteContent textarea:nth-child(3) {
-  height: 100px;
-}
-
-.question-box h3 {
-  color: #999999;
-  font-size: 20px;
-}
-.question-box input {
-  width: 100%;
-  background-color: transparent;
-  border: none;
-  border-bottom: 1px solid #999999;
-  padding: 5px;
-  margin-top: 10px;
-  color: white;
-}/*# sourceMappingURL=style.css.map */
