--- static/notePage/nav/style/style.css
@@ -1,16 +1,0 @@
-.nav {
-  padding: 10px;
-  cursor: pointer;
-  gap: 5px;
-}
-.nav img {
-  display: block;
-  width: 45px;
-  height: 45px;
-}
-.nav h1 {
-  margin: auto 0;
-  color: #2f403d;
-  font-size: 35px;
-  font-weight: 700;
-}/*# sourceMappingURL=style.css.map */
