--- static/notePage/sideBar/js/notePageSideBar.js
@@ -1,2 +1,0 @@
-function init() {}
-init();
