+++ static/notePage/style/main/ntMainToolBar.css
@@ -1,0 +1,79 @@
+main.edit-wrapper {
+  min-width: 800px;
+  flex-grow: 1;
+  padding-bottom: 20px;
+}
+
+.edit-area {
+  border-radius: 20px;
+  background-color: rgba(33, 33, 33, 0.7);
+  padding: 20px;
+}
+
+.edit-ctn-tools {
+  justify-content: space-between;
+}
+.edit-ctn-tools input {
+  flex-grow: 1;
+  padding-right: 10px;
+  background-color: transparent;
+  border: none;
+  color: white;
+  font-size: 24px;
+}
+
+.edit-group-tools {
+  gap: 5px;
+}
+.edit-group-tools img {
+  width: 35px;
+  height: 35px;
+  border: 1px solid #aaaaaa;
+  border-radius: 5px;
+  padding: 3px;
+  transition: all 0.2s ease-in-out;
+  cursor: pointer;
+}
+.edit-group-tools img:hover {
+  background-color: #aaaaaa;
+}
+.edit-group-tools #discard,
+.edit-group-tools #save {
+  font-size: 14px;
+  color: white;
+  border: 1px solid #aaaaaa;
+  border-radius: 5px;
+  margin: auto;
+  padding: 10px 20px;
+  transition: all 0.2s ease-in-out;
+  cursor: pointer;
+}
+.edit-group-tools div:nth-child(3) {
+  background-color: transparent;
+}
+.edit-group-tools div:nth-child(3):hover {
+  background-color: red;
+}
+.edit-group-tools div:nth-child(4) {
+  background-color: #263935;
+}
+.edit-group-tools div:nth-child(4):hover {
+  background-color: #3b5953;
+}
+.edit-group-tools .tagBtn {
+  position: relative;
+}
+.edit-group-tools .tagBtn .tagList {
+  position: absolute;
+  top: 35px;
+  left: 0px;
+}
+.edit-group-tools .tagBtn #done {
+  display: block;
+  margin-top: 5px;
+  position: static;
+  margin-left: auto;
+  background-color: green;
+  border-radius: 5px;
+  color: white;
+}/*# sourceMappingURL=ntMainToolBar.css.map */
