+++ static/notePage/style/main/writingArea.scss
