+++ static/notePage/style/nav/notePageNav.scss
