+++ static/notePage/style/sideBar/NotesSideBar.css
@@ -1,0 +1,140 @@
+.sideBar-ctn-notes {
+  padding: 10px;
+  background-color: rgba(33, 33, 33, 0.7);
+  border-radius: 10px;
+  -webkit-user-select: none;
+     -moz-user-select: none;
+          user-select: none;
+  height: calc(100vh - 425px);
+  overflow-y: scroll;
+}
+.sideBar-ctn-notes::-webkit-scrollbar {
+  display: none;
+  border: 1px solid white;
+}
+
+.notes-topToolsBar {
+  justify-content: space-between;
+  align-items: center;
+}
+.notes-topToolsBar p {
+  color: #cccccc;
+  font-weight: 500;
+}
+.notes-topToolsBar .tools {
+  gap: 5px;
+}
+.notes-topToolsBar .tools img {
+  border: 1px solid #aaaaaa;
+  border-radius: 5px;
+  padding: 3px;
+  width: 28px;
+  height: 28px;
+  transition: all 0.2s ease-in-out;
+  cursor: pointer;
+}
+.notes-topToolsBar .tools img:hover {
+  background-color: #aaaaaa;
+}
+
+.tagBtn {
+  position: relative;
+}
+
+.tagList {
+  position: absolute;
+  right: 0px;
+  top: 25px;
+  width: 180px;
+  background-color: #333333;
+  border: 1px solid #666666;
+  border-radius: 10px;
+  z-index: 20;
+  padding: 15px;
+  flex-direction: column;
+  gap: 10px;
+  cursor: default;
+}
+.tagList .tagItem {
+  justify-content: space-between;
+  gap: 5px;
+  align-items: center;
+}
+.tagList .tagItem p {
+  font-size: 14px;
+  border: 1px solid white;
+  border-radius: 5px;
+  color: white;
+  flex-grow: 1;
+  text-align: center;
+  padding: 5px;
+  transition: all 0.2s ease-in-out;
+  cursor: pointer;
+}
+.tagList .tagItem p:hover {
+  background-color: #666666;
+}
+.tagList .tagItem img {
+  width: 25px;
+  height: 25px;
+  flex-grow: 0;
+  cursor: pointer;
+}
+.tagList .tagBtnGroup {
+  margin-top: 10px;
+}
+.tagList .tagBtnGroup input {
+  width: 150px;
+  background-color: transparent;
+  border: none;
+  color: white;
+  border: none;
+  border-bottom: #666666 1px solid;
+  padding: 5px 50px 5px 5px;
+}
+.tagList .tagBtnGroup input::-moz-placeholder {
+  color: #999999;
+}
+.tagList .tagBtnGroup input::placeholder {
+  color: #999999;
+}
+.tagList .tagBtnGroup button {
+  position: absolute;
+  right: 15px;
+  border-radius: 5px;
+  background-color: #3d8344;
+  color: white;
+  cursor: pointer;
+  font-size: 14px;
+}
+
+.notes-group {
+  flex-direction: column;
+  gap: 10px;
+  margin-top: 10px;
+  overflow-y: scroll;
+}
+.notes-group::-webkit-scrollbar {
+  display: none;
+}
+
+.note-item {
+  border: 1px solid #aaaaaa;
+  border-radius: 8px;
+  padding: 8px 5px 8px 10px;
+  justify-content: space-between;
+  cursor: pointer;
+  transition: all 0.2s ease-in-out;
+}
+.note-item:hover {
+  background-color: #999999;
+}
+.note-item p {
+  font-size: 14px;
+  color: #cccccc;
+  margin: auto 0;
+}
+.note-item img {
+  width: 15px;
+  height: 15px;
+}/*# sourceMappingURL=NotesSideBar.css.map */
