+++ static/notePage/style/sideBar/collaborators.scss
