+++ static/notePage/style/sideBar/inviteCollaboratorBar.css
@@ -1,0 +1,53 @@
+.sideBar-wrapper {
+  flex-direction: column;
+  flex-grow: 0;
+  gap: 10px;
+  width: 250px;
+  height: calc(100vh - 65px);
+}
+
+.sideBar-ctn-invitePartner {
+  padding: 15px 20px;
+  background-color: rgba(33, 33, 33, 0.7);
+  border-radius: 10px;
+  -webkit-user-select: none;
+     -moz-user-select: none;
+          user-select: none;
+}
+.sideBar-ctn-invitePartner p {
+  color: #cccccc;
+}
+
+.inviteInput {
+  position: relative;
+  margin-top: 15px;
+}
+.inviteInput img {
+  position: absolute;
+  right: 0;
+  top: -8px;
+  padding: 3px;
+  border: #aaaaaa 1px solid;
+  border-radius: 5px;
+  width: 25px;
+  height: 25px;
+  transition: all 0.2s ease-in-out;
+  cursor: pointer;
+}
+.inviteInput img:hover {
+  background-color: #aaaaaa;
+}
+.inviteInput input {
+  width: 100%;
+  background-color: transparent;
+  border: none;
+  border-bottom: 1px solid #aaaaaa;
+  padding: 3px 30px 3px 3px;
+  color: white;
+}
+.inviteInput input::-moz-placeholder {
+  color: #aaaaaa;
+}
+.inviteInput input::placeholder {
+  color: #aaaaaa;
+}/*# sourceMappingURL=inviteCollaboratorBar.css.map */
