+++ static/notePage/style/sideBar/inviteCollaboratorBar.scss
