+++ static/notePage/style/sideBar/notesSideBar.scss
