--- static/userSpace/style.css
@@ -1,4 +1,0 @@
-@import url(/static/base/style.css);
-@import url(/static/userSpace/createNotebook/style/style.css);
-@import url(/static/userSpace/noteConsole/style/style.css);
-@import url(/static/userSpace/sideBar/style/style.css);
