+++ static/userSpace/style/createNotebookForm/createNotebookForm.css
@@ -1,0 +1,211 @@
+.main-wrapper {
+  width: calc(100vw - 270px);
+  position: relative;
+  gap: 30px;
+  overflow: hidden;
+}
+
+.toForm {
+  transition: all 0.3s ease-in-out;
+  transform: translate(calc(280px - 100vw));
+}
+
+.toNote {
+  transition: all 0.3s ease-in-out;
+  transform: translate(0);
+}
+
+.main-area-createNotebook {
+  width: 100%;
+  flex-shrink: 0;
+  margin: auto;
+}
+
+.main-ctn-createNotebook {
+  width: 500px;
+  height: 600px;
+  padding: 30px 20px 20px 20px;
+  margin: auto;
+  background-color: rgba(36, 49, 45, 0.6);
+  border-radius: 20px;
+  position: relative;
+  flex-shrink: 0;
+}
+.main-ctn-createNotebook p {
+  font-size: 24px;
+  color: white;
+  text-align: center;
+}
+
+.main-area-createNotebook form {
+  width: 460px;
+  padding: 20px 30px;
+  transition: all 0.3s ease-in-out;
+  gap: 30px;
+  overflow: hidden;
+}
+.main-area-createNotebook form label {
+  color: white;
+}
+.main-area-createNotebook form input {
+  border: none;
+  border-bottom: 1px solid #666666;
+  background: transparent;
+  padding: 0px 10px 7px 10px;
+  color: white;
+}
+.main-area-createNotebook form input::-moz-placeholder {
+  color: #aaaaaa;
+}
+.main-area-createNotebook form input::placeholder {
+  color: #aaaaaa;
+}
+
+.createNotebookPage {
+  width: 100%;
+  flex-direction: column;
+  flex-shrink: 0;
+  gap: 15px;
+  position: relative;
+}
+
+.addTag {
+  position: relative;
+}
+.addTag #addBtn {
+  padding: 7px;
+  border-radius: 5px;
+  font-size: 16px;
+  background-color: #1a491a;
+  position: absolute;
+  top: -5px;
+  right: 0;
+}
+.addTag input {
+  width: 100%;
+}
+
+.tags {
+  width: 100%;
+  gap: 10px;
+}
+.tags p {
+  padding: 7px;
+  border-radius: 5px;
+  font-size: 16px;
+  background-color: #1a491a;
+  cursor: pointer;
+  transition: all 0.1s ease-in-out;
+}
+.tags p:hover {
+  opacity: 0.8;
+}
+
+.description {
+  margin-top: 10px;
+  flex-direction: column;
+  gap: 10px;
+}
+.description textarea {
+  width: 100%;
+  height: 180px;
+  padding: 10px 20px;
+  border-radius: 10px;
+  background-color: transparent;
+  color: white;
+  resize: none;
+}
+
+.createNotebookFormbottomBtn {
+  justify-content: space-between;
+}
+.createNotebookFormbottomBtn div,
+.createNotebookFormbottomBtn button {
+  color: white;
+  padding: 7px 15px;
+  border-radius: 10px;
+  cursor: pointer;
+  transition: all 0.1s ease-in-out;
+}
+.createNotebookFormbottomBtn div:hover,
+.createNotebookFormbottomBtn button:hover {
+  opacity: 0.8;
+}
+.createNotebookFormbottomBtn div:nth-child(1) {
+  background-color: #666666;
+  margin-block: auto;
+}
+.createNotebookFormbottomBtn :nth-child(2) {
+  background-color: green;
+  border: none;
+}
+
+.inviteInput {
+  position: relative;
+}
+.inviteInput #addEmailBtn {
+  padding: 7px;
+  border-radius: 5px;
+  font-size: 16px;
+  background-color: #1a491a;
+  position: absolute;
+  top: -5px;
+  right: 0;
+  cursor: pointer;
+  transition: all 0.1s ease-in-out;
+}
+.inviteInput #addEmailBtn:hover {
+  opacity: 0.8;
+}
+.inviteInput input {
+  width: 100%;
+}
+
+.addEmailList {
+  flex-direction: column;
+  gap: 8px;
+}
+
+.email-item {
+  position: relative;
+}
+.email-item img {
+  position: absolute;
+  right: 20px;
+  bottom: 0;
+  width: 25px;
+  height: 25px;
+  cursor: pointer;
+}
+.email-item p {
+  font-size: 20px;
+  background-color: #1a491a;
+  padding: 7px 10px;
+  border-radius: 5px;
+  max-width: 320px;
+  overflow: hidden;
+  -webkit-user-select: none;
+     -moz-user-select: none;
+          user-select: none;
+  flex-grow: 0;
+}
+
+#errorMsg {
+  margin-inline: auto;
+  color: red;
+}
+
+#createFormCtn {
+  position: relative;
+  inset: 0 0 -30px 0;
+}
+
+.pageTwo {
+  transition: all 0.3s ease-in-out;
+  transform: translate(-440px);
+}
+
+.pageOne {
+  transition: all 0.3s ease-in-out;
+  transform: translate(0);
+}/*# sourceMappingURL=createNotebookForm.css.map */
