+++ static/userSpace/style/createNotebookForm/createNotebookForm.scss
