+++ static/userSpace/style/notebookSideBar/notebookSideBar.scss
