+++ static/userSpace/style/notesConsole/notesConsole.css
@@ -1,0 +1,240 @@
+.main-area-noteConsole {
+  width: calc(100% - 40px);
+  background-color: rgba(42, 42, 42, 0.8);
+  border-radius: 20px;
+  padding: 30px 20px;
+  flex-shrink: 0;
+  margin: 20px 0 20px 20px;
+}
+
+.main-ctn-topToolsBar {
+  justify-content: space-between;
+}
+.main-ctn-topToolsBar h2 {
+  color: white;
+  font-size: 24px;
+  font-weight: 500;
+}
+
+.main-group-tools {
+  gap: 5px;
+  -webkit-user-select: none;
+     -moz-user-select: none;
+          user-select: none;
+}
+
+.main-item-toolsBtn {
+  width: 38px;
+  height: 38px;
+  padding: 5px;
+  border: 1px solid #666666;
+  border-radius: 5px;
+  transition: all 0.2s ease-in-out;
+  cursor: pointer;
+}
+.main-item-toolsBtn:hover {
+  background-color: #666666;
+}
+.main-item-toolsBtn img {
+  width: 100%;
+  height: 100%;
+  vertical-align: bottom;
+}
+
+.tagList {
+  position: relative;
+  left: -5px;
+  top: 5px;
+  width: 220px;
+  background-color: #333333;
+  border: 1px solid #666666;
+  border-radius: 10px;
+  z-index: 15;
+  padding: 15px;
+  flex-direction: column;
+  gap: 10px;
+  cursor: default;
+}
+.tagList .tagItem {
+  justify-content: space-between;
+  gap: 5px;
+  align-items: center;
+}
+.tagList .tagItem p {
+  border: 1px solid white;
+  border-radius: 5px;
+  color: white;
+  flex-grow: 1;
+  text-align: center;
+  padding: 5px;
+  transition: all 0.2s ease-in-out;
+  cursor: pointer;
+}
+.tagList .tagItem p:hover {
+  background-color: #666666;
+}
+.tagList .tagItem img {
+  width: 25px;
+  height: 25px;
+  flex-grow: 0;
+  cursor: pointer;
+}
+.tagList .tagBtnGroup {
+  margin-top: 10px;
+}
+.tagList .tagBtnGroup input {
+  width: 190px;
+  background-color: transparent;
+  border: none;
+  color: white;
+  border: none;
+  border-bottom: #666666 1px solid;
+  padding: 5px 50px 5px 5px;
+}
+.tagList .tagBtnGroup input::-moz-placeholder {
+  color: #999999;
+}
+.tagList .tagBtnGroup input::placeholder {
+  color: #999999;
+}
+.tagList .tagBtnGroup button {
+  position: absolute;
+  right: 15px;
+  border-radius: 5px;
+  background-color: #3d8344;
+  color: white;
+  cursor: pointer;
+}
+
+.main-item-toolsSearchBar {
+  gap: 15px;
+  position: relative;
+}
+.main-item-toolsSearchBar input {
+  width: 200px;
+  background-color: transparent;
+  border: 1px solid #666666;
+  border-radius: 5px;
+  font-size: 16px;
+  padding: 6px 35px 6px 15px;
+  color: white;
+}
+.main-item-toolsSearchBar img {
+  width: 25px;
+  height: 25px;
+  position: absolute;
+  right: 7px;
+  bottom: 7px;
+  cursor: pointer;
+}
+
+.main-ctn-noteConsole {
+  width: calc(100% - 40px);
+  height: calc(100% - 40px);
+  background-color: rgba(41, 53, 45, 0.5);
+  border-radius: 20px;
+  margin: 10px auto 0 auto;
+  padding: 20px;
+  position: relative;
+}
+
+.newNoteBtn {
+  position: absolute;
+  right: 20px;
+  width: 100px;
+  color: white;
+  background-color: #3d8344;
+  text-align: center;
+  border-radius: 10px;
+  padding: 8px 5px;
+}
+
+.main-group-subjectInfo {
+  color: white;
+}
+.main-group-subjectInfo h2 {
+  font-size: 24px;
+  padding: 10px 0;
+}
+.main-group-subjectInfo p {
+  width: 85%;
+  padding: 10px;
+}
+
+.main-group-notes {
+  display: grid;
+  gap: 10px;
+  margin-top: 10px;
+  height: calc(100vh - 300px);
+  border-radius: 20px;
+  overflow-y: scroll;
+}
+.main-group-notes::-webkit-scrollbar {
+  display: none;
+}
+@media (min-width: 1500px) {
+  .main-group-notes {
+    grid-template-columns: repeat(5, 1fr);
+  }
+}
+@media (min-width: 1350px) and (max-width: 1500px) {
+  .main-group-notes {
+    grid-template-columns: repeat(4, 1fr);
+  }
+}
+@media (min-width: 1200px) and (max-width: 1350px) {
+  .main-group-notes {
+    grid-template-columns: repeat(3, 1fr);
+  }
+}
+@media (max-width: 1200px) {
+  .main-group-notes {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+@media (max-width: 960px) {
+  .main-group-notes {
+    grid-template-columns: repeat(1, 1fr);
+  }
+}
+
+.main-item-noteCard {
+  background-color: #666666;
+  border-radius: 15px;
+  padding: 10px 15px 35px 15px;
+  position: relative;
+  color: white;
+  flex-direction: column;
+  gap: 10px;
+  transition: all 0.2s ease-in-out;
+  cursor: pointer;
+}
+.main-item-noteCard:hover {
+  opacity: 0.8;
+}
+.main-item-noteCard h3 {
+  font-size: 22px;
+  padding-left: 2px;
+}
+.main-item-noteCard p {
+  margin-top: 5px;
+  font-size: 16px;
+  overflow: hidden;
+  transition: all 0.1s ease-in-out;
+}
+.main-item-noteCard p::-webkit-scrollbar {
+  display: none;
+}
+.main-item-noteCard img {
+  position: absolute;
+  right: 8px;
+  top: 8px;
+  width: 25px;
+  height: 25px;
+}
+
+.date {
+  position: absolute;
+  bottom: 10px;
+  right: 10px;
+}/*# sourceMappingURL=notesConsole.css.map */
