+++ static/userSpace/style/notesConsole/notesConsole.scss
