+++ static/userSpace/style/style.css
@@ -1,0 +1,4 @@
+@import url(/static/base/style.css);
+@import url(/static/userSpace/style/createNotebookForm/createNotebookForm.css);
+@import url(/static/userSpace/style/notesConsole/notesConsole.css);
+@import url(/static/userSpace/style/notebookSideBar/notebookSideBar.css);
