+++ userSpace.html
@@ -13,7 +13,7 @@
     />
     <!-- /google font/ -->
     <link rel="icon" href="/static/resource/images/nodes-blue.png" />
-    <link rel="stylesheet" href="/static/userSpace/style.css" />
+    <link rel="stylesheet" href="/static/userSpace/style/style.css" />
     <!-- /shared resuroce/ -->
     <title>userSpace</title>
   </head>
@@ -1390,8 +1390,8 @@
       </main>
     </div>
     <script src="/static/base/js/base.js"></script>
-    <script src="/static/userSpace/sideBar/js/userSpaceSidebar.js"></script>
-    <script src="/static/userSpace/createNotebook/js/userSpaceCreateNotbook.js"></script>
-    <script src="/static/userSpace/noteConsole/js/userSpaceNoteConsole.js"></script>
+    <script src="/static/userSpace/js/userSpaceSidebar.js"></script>
+    <script src="/static/userSpace/js/userSpaceCreateNotbook.js"></script>
+    <script src="/static/userSpace/js/userSpaceNoteConsole.js"></script>
   </body>
 </html>
